requirements.txt
beautifulsoup4>=4.6.3
clipboard>=0.0.4
elevate>=0.1.3
keyring>=20.0.0
keyrings.cryptfile>=1.3.8
playsound>=1.2.2
psutil>=5.0.0
requests>=2.15.0
rich>=10.0.1
selenium==3.141.0
webdriver-manager>=3.0.0
